Fix the defects -Wall was hiding
Second of four staged PRs for #138: everything -Wall reports that needed
individual reasoning rather than a mechanical edit. 251 -> 220 warnings.

Real defects:

  * httpdsl.c's "must be called by the HTTPD web server" diagnostic had two
    %s and one argument, in both the wtof() and the printf(). The second
    conversion read the varargs area past the end and dereferenced it as a
    char *, so the message meant to explain a TSO misuse could abend
    instead. The four sibling CGIs (httpdm, httpdmtt, httpdsrv, httpjes2)
    all pass argv[0] here; httpdsl was the one that lost it.

  * jesst.c closed its "dd" JSON array with printf("%s]\n") and no argument
    at all -- same shape. Every other line in that function passes the
    indent string in, which is what belongs there.

  * dbgdump.c printed a pointer through %08X and dbgtime.c a long through
    %06u.

Implicit declarations, 13 of them. Four in httpd.c and one in credtest.c
just needed clibauth.h; httpstrt.c needed time.h, clibenv.h and clibwto.h.
Two were calls to CSECT names that have no C prototype: httpfcgi.c called
httpcmp() and httppcgi.c called httplink(), which are HTTPCMP and HTTPLINK
-- the declared spellings are http_cmp/http_link, so those files now #undef
the httpx macro and call the direct extern, the same pattern httpxauth.c
and httpgufs.c already use. httpdslp.c called __arcou(), which is not an
API name at all: the declared entry for @@ARCOU is array_count().

sleep() and __tzset() are the exception -- libc370 ships both but declares
neither in any header, so httpprm.c declares them locally with the
signatures copied from the libc370 definitions. Filed against libc370
separately; a later header will agree rather than conflict.

httpopen.c's mode joins the false positives: it is only assigned inside
if (mime), but http_mime() never returns NULL -- an extension it does not
know falls back to default_mime {"txt","text/plain",0}, which is exactly
the "r" the initializer now spells out. Verified on MVS: a .zzq file serves
200 text/plain. GCC cannot see through the external call, hence the warning.

The remaining three are false positives silenced defensively: progLen in
httpstrt.c/cgistart.c is written and read under the same TSO flag, which
the compiler cannot correlate, and httpcred.c's rc is only unwritten if the
static head[] array were empty.

Also: httpx.c had a nested /* inside a disabled vector entry's comment, and
httpdsld.c initialized two structs whose first member is an aggregate with
{0} instead of {{0}}. Neither changes the vector layout or the zeroing.

The 18 subscript-has-type-char warnings are deliberately NOT here: on
EBCDIC with signed char a set high bit indexes negatively, and the fix is
only correct after checking each table's declared size, so they get their
own PR rather than a blanket cast.

Refs #138

+++ src/dbgdump.c
@@ -22,7 +22,7 @@ dump(FILE *fp, const char *pName, const void *pvArea, int iSize, int iChunk)
 
     fprintf(fp, "%*.*sDump of %08X \"%s\" (%d bytes)\n",
         indent, indent, "",
-        pArea, pName, iSize);
+        (unsigned int)pArea, pName, iSize);
 
     for (i=0, j=0; iSize > 0;i++ ) {
         if ( i==iChunk ) {
